Finished analysist of stock portfolios
This file will take you through how I analyzed stocks that were given to me, as well as a few hand picked ones. I use Python and its environments to output  statistics to determine which portfolios were admirable. Getting number like:
- Sharpe ratios
- Rolling sharpe ratios 
- Beta
- Rolling beta
- Daily returns
- Annual cumulative returns
- etc..

Throughout my coding I have included what my processes were so it is easier to keep track. This is my first time doing coding at this scale so please excuse the disorganization of it and hopefully it is precise enough to read and comprehend.